Restructured Util tests
